src/crawler/requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
fastapi==0.111.0
uvicorn[standard]==0.30.1
transformers==4.41.2
torch==2.3.1
requests==2.31.0
python-dotenv==1.0.1
urllib3==2.0.7
certifi==2023.11.17
pyodbc==5.1.0
pydantic==2.7.1
# Celery task queue
celery[redis]==5.4.0
kombu==5.3.7